requirements.txt
Django>=4.2,<5.0
requests
feedparser
python-dateutil
gunicorn
whitenoise